+++ frontend/src/components/Dashboard/ServiceEvents.vue
@@ -1,28 +1,31 @@
 <template>
   <div>
     <Loading :loading="!loaded"/>
-    <div v-if="loaded && !service.online" class="bg-white shadow-sm mt-3 p-3 pr-4 pl-4 col-12">
-      <font-awesome-icon icon="exclamation" class="mr-3" size="1x"/>
-      Last failure was {{ago(service.last_error)}} ago.
-      <code v-if="failure" class="d-block bg-light p-3 mt-3">
-        {{failure.issue}}
-        <span class="d-block text-dim float-right small mt-3 mb-1">Failure #{{failure.id}}</span>
-      </code>
-    </div>
-    <div v-if="loaded" v-for="message in $store.getters.serviceMessages(service.id)" class="bg-light shadow-sm p-3 pr-4 pl-4 col-12 mt-3">
-      <font-awesome-icon icon="calendar" class="mr-3" size="1x"/> {{message.description}}
-      <span class="d-block small text-muted mt-3">
-        Starts at <strong>{{niceDate(message.start_on)}}</strong> till <strong>{{niceDate(message.end_on)}}</strong>
-        ({{dur(parseISO(message.start_on), parseISO(message.end_on))}})
+    <div v-if="loaded && isBefore(nowSubtract(86400), service.last_error)" class="text-danger font-2 p-0 m-0 mb-2">
+      <font-awesome-icon icon="exclamation" class="mr-1 text-danger" size="1x"/> Recent Failure<br>
+      <span class="font-italic d-inline-block text-dim" style="max-width: 270px">
+      Last failure was {{ago(service.last_error)}} ago. {{service.failures[0].issue}}
       </span>
     </div>
-    <div v-if="loaded" v-for="incident in incidents" class="bg-light shadow-sm p-3 pr-4 pl-4 col-12 mt-3">
-      <font-awesome-icon icon="calendar" class="mr-3" size="1x"/>
-      {{incident.title}} - {{incident.description}}
-      <div v-for="update in incident.updates" class="d-block small">
-        <span class="font-weight-bold text-capitalize">{{update.type}}</span> - {{update.message}}
-      </div>
+
+
+    <div v-if="loaded" v-for="message in messages" class="font-2 p-0 m-0 mb-2">
+      <font-awesome-icon icon="calendar" class="mr-1" size="1x"/> Upcoming Announcement<br>
+      <span class="font-italic font-weight-light">{{message.description}}</span>
+      <span class="font-0 text-dim float-right font-weight-light">@ <strong>{{niceDate(message.start_on)}}</strong>
+      </span>
     </div>
+    <div v-if="loaded" v-for="incident in incidents" class="font-2 p-0 m-0 mb-2">
+      <font-awesome-icon icon="bullhorn" class="mr-1" size="1x"/>Recent Incident<br>
+      <span class="font-italic d-inline-block text-truncate" style="max-width: 270px">{{incident.title}} - {{incident.description}}</span>
+      <span class="font-0 text-dim float-right font-weight-light">@ <strong>{{niceDate(incident.created_at)}}</strong></span>
+    </div>
+
+    <div v-if="success_event" class="font-2 p-0 m-0 mt-1 mb-3">
+      <span class="text-success"><font-awesome-icon icon="check" class="mr-1" size="1x"/>No New Events</span>
+      <span class="font-italic d-inline-block text-truncate text-dim" style="max-width: 270px">Last failure was {{ago(service.last_error)}} ago.</span>
+    </div>
+
   </div>
 </template>
 
@@ -34,7 +37,6 @@ export default {
 name: "ServiceEvents",
   components: {
     Loading
-
   },
   props: {
     service: {
@@ -55,13 +57,19 @@ name: "ServiceEvents",
   computed: {
     messages() {
       return this.$store.getters.serviceMessages(this.service.id)
+    },
+    success_event() {
+      if (this.service.online && this.service.messages.length === 0 && this.service.incidents.length === 0) {
+        return true
+      }
+      return false
     }
   },
   methods: {
     async load() {
       this.loaded = false
       if (!this.service.online) {
-        await this.getFailure()
+        this.failure = this.service.failures[0]
       }
       await this.getMessages()
       await this.getIncidents()
@@ -69,10 +77,6 @@ name: "ServiceEvents",
     },
     async getMessages() {
       // this.messages = this.$store.getters.serviceMessages(this.service.id)
-    },
-    async getFailure() {
-      const f = await Api.service_failures(this.service.id, null, null, 1)
-      this.failure = f[0]
     },
     async getIncidents() {
       this.incidents = await Api.incidents_service(this.service.id)