+++ frontend/src/pages/Help.vue
@@ -2275,7 +2275,7 @@ OluFxewsEO0QNDrfFb+0gnjYlnGqOFcZjUMXbDdY5oLSPtXohynuTK1qyQ==
 </div>
 
 <div class="text-center small text-dim" v-pre>
-Automatically generated from Statping's Wiki on 2020-08-20 20:45:46.398081 &#43;0000 UTC
+Automatically generated from Statping's Wiki on 2020-08-21 04:07:11.080469 &#43;0000 UTC
 </div>
 
 </div>