+++ handlers/index.go
@@ -22,8 +22,3 @@ func healthCheckHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	returnJson(health, w, r)
 }
-
-func notFoundHandler(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusNotFound)
-	ExecuteResponse(w, r, "base.gohtml", core.App, nil)
-}